add message state in tera module
