+++ magic_bridge/ic/src/dip20_proxy/src/lib.rs
@@ -11,9 +11,6 @@ fn main() {}
 #[cfg(not(any(target_arch = "wasm32", test)))]
 fn main() {
     use ic_kit::candid;
-    use ic_kit::candid::Nat;
-    use ic_kit::Principal;
-    use types::*;
 
     candid::export_service!();
     std::print!("{}", __export_service());
